+++ projects/_04_direct_payment_recipients/fargate/utils/estimate_direct_payments_utils/calculate_remaining_variables.py
@@ -3,7 +3,7 @@ import polars as pl
 from projects._04_direct_payment_recipients.direct_payments_column_names import (
     DirectPaymentColumnNames as DP,
 )
-from projects._04_direct_payment_recipients.direct_payments_configuration import (
+from projects._04_direct_payment_recipients.direct_payments_config_polars import (
     DirectPaymentConfiguration as Config,
 )